ircommander_client/requirements.txt
# iRCommander Client
# A lightweight PC service for iRacing telemetry and control

# HTTP client (still needed for some operations)
requests>=2.28.0

# Supabase client
supabase>=2.0.0

# Environment configuration
python-dotenv>=1.0.0

# iRacing SDK
pyirsdk>=1.3.0

# GUI (PyQt6)
PyQt6>=6.4.0

# System information
psutil>=5.9.0

# Development tools
watchdog>=3.0.0

# Build tools (optional - for creating executables)
pyinstaller>=5.0

# Windows autostart utility (optional - only needed for Python script method)
# pywin32>=305

# Remote Desktop (WebRTC)
aiortc>=1.6.0
opencv-python>=4.8.0
mss>=9.0.0
numpy>=1.24.0
aiohttp>=3.9.0

# Input simulation for remote control
pyautogui>=0.9.54

